KTOR-973 Make pipeline frozen
