+++ ktor-utils/posix/src/io/ktor/util/AttributesNative.kt
@@ -4,13 +4,15 @@
 
 package io.ktor.util
 
+import io.ktor.util.collections.*
+
 /**
  * Create native specific attributes instance.
  */
 public actual fun Attributes(concurrent: Boolean): Attributes = AttributesNative()
 
 private class AttributesNative : Attributes {
-    private val map = mutableMapOf<AttributeKey<*>, Any?>()
+    private val map = ConcurrentMap<AttributeKey<*>, Any>()
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : Any> getOrNull(key: AttributeKey<T>): T? = map[key] as T?
