+++ ktor-utils/posix/src/io/ktor/util/CoroutinesUtilsNative.kt
@@ -5,6 +5,7 @@
 package io.ktor.util
 
 import kotlin.coroutines.*
+import kotlin.coroutines.intrinsics.*
 
 @InternalAPI
 @Suppress("NOTHING_TO_INLINE")
@@ -14,9 +15,7 @@ internal actual inline fun <R, A>
     arg: A,
     continuation: Continuation<Unit>
 ): Any? {
-
-    @Suppress("UNCHECKED_CAST")
-    val function = (this as Function3<R, A, Continuation<Unit>, Any?>)
-    return function.invoke(receiver, arg, continuation)
+    val block: suspend () -> Unit = { this(receiver, arg) }
+    return block.startCoroutineUninterceptedOrReturn(continuation)
 }
